moved statblock to entity superclass
